dedup setImage
